requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.30
insightface==0.7.3
numpy==1.26.4
scipy==1.12.0
flask==3.0.2
sqlalchemy==2.0.27
pycryptodome==3.20.0
Pillow==10.2.0
onnxruntime==1.17.1
dlib==19.24.2
imutils==0.5.4